requirements.txt
# Web Framework
Flask==2.0.1
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SocketIO==5.1.1
gunicorn==21.2.0

# Veritabanı
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
SQLAlchemy==1.4.23

# Asenkron işler
eventlet==0.33.3

# Yapay Zeka ve Görüntü İşleme
tensorflow==2.14.0
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56
Pillow==8.3.2
numpy==1.21.2
scikit-learn==0.24.2
matplotlib==3.4.3

# Pytorch-Age-Estimation modeli için
facenet-pytorch==2.5.2
timm>=1.0.0

# Nesne Tespit ve Yüz Analizi
ultralytics==8.0.0  # YOLO
transformers==4.31.0  # CLIP
# dlib kaldırıldı çünkü yerini MTCNN aldı
face-recognition==1.3.0  # Yüz tanıma

# Dosya İşleme
python-magic==0.4.27  # MIME türü algılama
moviepy==1.0.3  # Video işleme

# Güvenlik
python-dotenv==0.19.0
Flask-JWT-Extended==4.5.2
bcrypt==4.0.1

# Test ve Geliştirme
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.1.0

# Ek Gereksinimler
gdown
requests 